+++ examples/12-graphql-apollo/src/subscriptions/subscription.providers.ts
@@ -7,9 +7,7 @@ export const createSubscriptionProviders = (port: number = 3001) => [
     provide: SUBSCRIPTION_SERVER,
     useFactory: () => {
       const server = createServer();
-      return new Promise(resolve =>
-        server.listen(port, () => resolve(server)),
-      );
+      return new Promise(resolve => server.listen(port, () => resolve(server)));
     },
   },
 ];