+++ lib/core/injector/module.d.ts
@@ -54,7 +54,12 @@ export declare class Module {
     addCustomFactory(component: CustomFactory, collection: Map<string, any>): void;
     addExportedComponent(exportedComponent: ComponentMetatype | string | DynamicModule): Set<string>;
     addCustomExportedComponent(exportedComponent: CustomFactory | CustomValue | CustomClass): Set<string>;
+<<<<<<< HEAD
     addRoute(route: Type<Controller>): void;
+=======
+    validateExportedProvider(token: string): string;
+    addRoute(route: Metatype<Controller>): void;
+>>>>>>> master
     addRelatedModule(relatedModule: any): void;
     replace(toReplace: any, options: any): string;
     createModuleRefMetatype(components: any): {