+++ src/common/interceptors/multer/multer.utils.ts
@@ -1,4 +1,9 @@
-import { InternalServerErrorException, HttpException, PayloadTooLargeException, BadRequestException } from './../../exceptions';
+import {
+  InternalServerErrorException,
+  HttpException,
+  PayloadTooLargeException,
+  BadRequestException,
+} from './../../exceptions';
 import { multerExceptions } from './multer.constants';
 
 export function transformException(error: Error | undefined) {
@@ -6,11 +11,11 @@ export function transformException(error: Error | undefined) {
     return error;
   }
   switch (error.message) {
-    case multerExceptions.LIMIT_FILE_SIZE: 
+    case multerExceptions.LIMIT_FILE_SIZE:
       return new PayloadTooLargeException(error.message);
     case multerExceptions.LIMIT_FILE_COUNT:
-    case multerExceptions.LIMIT_FIELD_KEY: 
-    case multerExceptions.LIMIT_FIELD_VALUE: 
+    case multerExceptions.LIMIT_FIELD_KEY:
+    case multerExceptions.LIMIT_FIELD_VALUE:
     case multerExceptions.LIMIT_FIELD_COUNT:
     case multerExceptions.LIMIT_UNEXPECTED_FILE:
     case multerExceptions.LIMIT_PART_COUNT: