+++ src/core/test/router/routes-resolver.spec.ts
@@ -84,7 +84,10 @@ describe('RoutesResolver', () => {
       const spy = sinon
         .stub(routesResolver, 'registerRouters')
         .callsFake(() => undefined);
-      routesResolver.resolve({ use: sinon.spy() } as any, { use: sinon.spy() } as any);
+      routesResolver.resolve(
+        { use: sinon.spy() } as any,
+        { use: sinon.spy() } as any,
+      );
       expect(spy.calledTwice).to.be.true;
     });
   });