+++ src/microservices/test/server/server-tcp.spec.ts
@@ -60,7 +60,7 @@ describe('ServerTCP', () => {
       expect(
         socket.sendMessage.calledWith({
           status: 'error',
-          error: NO_PATTERN_MESSAGE,
+          err: NO_PATTERN_MESSAGE,
         }),
       ).to.be.true;
     });